Serialização de NFS-e sem xsd

+++ pynfe/entidades/servico.py
@@ -8,7 +8,7 @@ from decimal import Decimal
 class Servico(Entidade):
 
     valor_servico = Decimal()
-    iss_retido = Decimal()
+    iss_retido = int()
     """ http://www1.receita.fazenda.gov.br/sistemas/nfse/tabelas-de-codigos.htm 
         Lista com códigos dos serviços
     """
